ByteToMessageDecoder$CompactingAccumulator
Motivation:
ByteToMessageDecoder relies upon the Cumulator to accumulate data accross
read/decode calls and compacts the cumulation ByteBuf in an class
method discardSomeReadBytes(). The strategy used between these two operations
maybe coupled but currently it is not possible to override both. For example a
strategy for accumulation maybe to avoid modifying the underlying structure of
the ByteBuf (e.g. no implicit re-allocation and no in place compaction of data
to free up space) and another stategy is to treat the buffer as fully modifiable
given there is no shared usage (e.g. the current implementation).

Modifications:
- Add ByteToMessageDecoder$CompactingAccumulator which extends Cumulation and has
  control over how data is compacted.

Result:
ByteToMessageDecoder$CompactingAccumulator has control over how data is
accumulated and how it is compacted. Therefore it has control over how data
behaves in the cumulation ByteBuf.